Extracted composables to the components folder. Each screen will have its own components folder that hosts these composables. #1
Moved UI folder to the app folder.
