+++ app/src/main/java/com/sivan/sbaceman/presentation/home_screen/components/ParticipantCount.kt
@@ -1,0 +1,26 @@
+package com.sivan.sbaceman.presentation.home_screen.components
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.material.Icon
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.rounded.Person
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.sp
+
+@Composable
+fun ParticipantCount(modifier: Modifier) {
+    Row(modifier = modifier, horizontalArrangement = Arrangement.SpaceEvenly) {
+        Icon(imageVector = Icons.Rounded.Person, contentDescription = "Participant count")
+        Text(
+            text = "50",
+            style = MaterialTheme.typography.labelMedium.copy(
+                fontSize = 16.sp,
+                color = MaterialTheme.colorScheme.onSecondaryContainer
+            )
+        )
+    }
+}
