+++ app/src/main/java/com/sivan/sbaceman/presentation/home_screen/components/SearchBar.kt
@@ -1,0 +1,74 @@
+package com.sivan.sbaceman.presentation.home_screen.components
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.text.BasicTextField
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.focus.onFocusChanged
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun SearchBar(
+    modifier: Modifier,
+    hint: String,
+    onSearchButtonClicked: (String) -> Unit
+) {
+    var searchText by rememberSaveable { mutableStateOf("") }
+
+    var isHintDisplayed by remember {
+        mutableStateOf(hint != "")
+    }
+
+    Box(modifier = modifier) {
+        BasicTextField(
+            value = searchText,
+            onValueChange = {
+                searchText = it
+            },
+            singleLine = true,
+            textStyle = MaterialTheme.typography.bodyMedium,
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(horizontal = 20.dp, vertical = 16.dp)
+                .onFocusChanged {
+                    isHintDisplayed = !it.hasFocus
+                },
+        )
+
+        if (isHintDisplayed && searchText.isEmpty() || searchText.isEmpty()) {
+            Text(
+                text = hint,
+                modifier = Modifier
+                    .align(Alignment.CenterStart)
+                    .padding(horizontal = 20.dp),
+                style = MaterialTheme.typography.bodyMedium
+            )
+        }
+    }
+}
+
+@Composable
+@Preview
+fun SearchBarPreview() {
+    SearchBar(
+        modifier = Modifier
+            .fillMaxWidth()
+            .height(80.dp),
+        hint = "Search for a space",
+        onSearchButtonClicked = {
+        }
+    )
+}
