+++ app/src/main/java/com/sivan/sbaceman/presentation/home_screen/components/SpaceTimings.kt
@@ -1,0 +1,30 @@
+package com.sivan.sbaceman.presentation.home_screen.components
+
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.width
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+
+@Composable
+fun SpaceTimings(modifier: Modifier, time: String) {
+    Row(modifier = modifier) {
+        Text(
+            text = "Starts at",
+            fontSize = 16.sp,
+            color = MaterialTheme.colorScheme.onSecondaryContainer
+        )
+        Spacer(modifier = Modifier.width(4.dp))
+        Text(
+            text = time,
+            style = MaterialTheme.typography.labelMedium.copy(
+                fontSize = 16.sp,
+                color = MaterialTheme.colorScheme.onSecondaryContainer
+            )
+        )
+    }
+}
