+++ app/src/main/java/com/sivan/sbaceman/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.sivan.sbaceman.presentation.ui.theme
+package com.sivan.sbaceman.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
